# optimize `inflate::core::init_tree` by precomputing reversed bits