@oyvindln 2023-01-10
@notgull does it look fine to you as well now?