@notgull 2023-01-10
Yeah, it looks good to me. This is basically how I'd do it.