@oyvindln 2023-01-10
aight. let me know when you want a new release. Can wait a bit if you have more work on the way.